:recycle: move conversion commands to another folder
