+++ tests/conversion/dof/empty_mgr.dof
@@ -1,0 +1,12 @@
+qc2zx
+qc2ts
+zx2ts
+
+qcir new
+qc2zx
+qc2ts
+
+zx new
+zx2ts
+
+q -f
